Moved most of stctrl binary code to lib.
