+++ components/stctrl/src/lib.rs
@@ -22,5 +22,7 @@ pub mod config;
 pub mod funds;
 pub mod info;
 
+pub mod stctrllib;
+
 pub mod file;
 pub mod utils;
